cli/pyproject.toml
[project]
name = "dbt-forge"
dynamic = ["version"]
description = "Scaffold production-ready dbt projects with opinionated defaults"
readme = { file = "README.md", content-type = "text/markdown" }
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.11,<3.14"
authors = [
    { name = "Marouane" },
]
maintainers = [
    { name = "Marouane" },
]
keywords = [
    "analytics-engineering",
    "cli",
    "code-generator",
    "dbt",
    "scaffolding",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Utilities",
]
urls = { Homepage = "https://github.com/maroil/dbt-forge", Repository = "https://github.com/maroil/dbt-forge", Issues = "https://github.com/maroil/dbt-forge/issues", Changelog = "https://github.com/maroil/dbt-forge/blob/main/CHANGELOG.md" }
dependencies = [
    "jinja2>=3.1.6",
    "pyyaml>=6.0",
    "questionary>=2.1.1",
    "rich>=14.3.3",
    "typer>=0.12",
]

[project.optional-dependencies]
snowflake = ["snowflake-connector-python>=3.0"]
bigquery = ["google-cloud-bigquery>=3.0"]
postgres = ["psycopg2-binary>=2.9"]
duckdb = ["duckdb>=1.0"]
databricks = ["databricks-sql-connector>=3.0"]
redshift = ["psycopg2-binary>=2.9"]
trino = ["trino>=0.320"]
spark = ["pyhive[hive]>=0.7"]
claude = ["anthropic>=0.30"]
openai = ["openai>=1.30"]

[project.scripts]
dbt-forge = "dbt_forge.main:app"

[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/dbt_forge/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
    "/pyproject.toml",
]

[tool.hatch.build.targets.wheel]
packages = ["src/dbt_forge"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["integration: end-to-end tests requiring dbt + DuckDB"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
major_version_zero = true

[dependency-groups]
dev = [
    "build>=1.2.2",
    "commitizen>=4.9.1",
    "dbt-core>=1.8.0",
    "dbt-duckdb>=1.8.0",
    "duckdb>=1.0.0",
    "pre-commit>=4.3.0",
    "pytest>=9.0.2",
    "pytest-timeout>=2.3.1",
    "pyyaml>=6.0.3",
    "ruff>=0.13.0",
    "twine>=6.1.0",
]
